Put all synthetic events in one file
