--- packages/react-dom/src/events/SyntheticInputEvent.js
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import SyntheticEvent from '../events/SyntheticEvent';
-
-/**
- * @interface Event
- * @see http://www.w3.org/TR/2013/WD-DOM-Level-3-Events-20131105
- *      /#events-inputevents
- */
-const SyntheticInputEvent = SyntheticEvent.extend({
-  data: null,
-});
-
-export default SyntheticInputEvent;
