+++ packages/react-dom/src/events/plugins/ModernEnterLeaveEventPlugin.js
@@ -13,8 +13,7 @@ import {
   TOP_POINTER_OVER,
 } from '../DOMTopLevelEventTypes';
 import {IS_REPLAYED} from 'react-dom/src/events/EventSystemFlags';
-import SyntheticMouseEvent from '../SyntheticMouseEvent';
-import SyntheticPointerEvent from '../SyntheticPointerEvent';
+import {SyntheticMouseEvent, SyntheticPointerEvent} from '../SyntheticEvent';
 import {
   getClosestInstanceFromNode,
   getNodeFromInstance,
